Each thread now writes its own files
Before, each thread would send the etyms one at a time to the main
thread over a channel. The main thread would then write out each one. To
take the most advantage of concurrency, however, each thread now handles
file writing on its own.

This change also replaces the loop+match in the threads with a for loop,
because that's just cleaner. It also moves creation of the Core and
Client to lib.rs, so that only one of each needs to be made per thread.

Joining on all the threads is now necessary because program exit is no
longer blocked on the main thread's file writing.