Make page iterator only owned by page_sender thread
It only needed to be under a mutex because the worker threads were
setting the progress bar's length to the iterator's length at every
iteration, which honestly was never really necessary. It was only useful
in a design where the length of the iterator was determined as it
iterated, rather than during `initialize`. This design was one of
multiple options, but I never chose it and I guess the mutex thing got
left in the code accidentally.